spottedcow_viz/requirements.txt
functions-framework==3.*
pandas
matplotlib
psycopg2-binary
google-cloud-storage